Host: Accommodation implementation (CRUD) - WIP
