+++ resources/lang/ro.json
@@ -132,7 +132,7 @@
     "Edit note": "Editează notă",
     "Delete note": "Sterge notă",
     "Are you sure you want to delete this note": "Sigur vrei sa ștergi această notă",
-    "Are you sure you want to delete this accommodation": "Sigur vrei sa ștergi această cazare",
+    "Are you sure you want to delete this accommodation": "Sigur vrei sa ștergi cazarea selectată",
     "Cancel": "Renunță",
     "Yes": "Da",
     "No": "Nu",
