+++ resources/views/host/view-accommodation.blade.php
@@ -129,6 +129,14 @@
                     </div>
                 </div>
             </div>
+
+            <div class="row">
+                <div class="col-sm-6">
+                    <h5 class="font-weight-600 text-primary mb-4 mt-4">{{ __('Description') }}</h5>
+                    <p>{!! $accommodation->description !!}</p>
+                </div>
+            </div>
+
             <div class="border-top pt-3 mt-4">
                 <div class="row">
                     <div class="col-sm-6">
